вынос отдельных функций
